+++ src/main/java/com/catdog/help/web/form/itemboard/UpdateItemBoardForm.java
@@ -1,7 +1,6 @@
 package com.catdog.help.web.form.itemboard;
 
 import com.catdog.help.domain.board.ItemBoard;
-import com.catdog.help.domain.board.UploadFile;
 import com.catdog.help.web.form.uploadfile.ReadUploadFileForm;
 import lombok.Getter;
 import lombok.Setter;
@@ -44,7 +43,7 @@ public class UpdateItemBoardForm {
 
     private List<MultipartFile> newImages = new ArrayList<>();
 
-    private List<Integer> deleteImageIds = new ArrayList<>();
+    private List<Long> deleteImageIds = new ArrayList<>();
 
     public UpdateItemBoardForm() { //컨트롤러 파라미터 바인딩
     }
